:recycle: Refactor `NGFFWSIReader`
