+++ tiatoolbox/wsicore/wsireader/__init__.py
@@ -9,7 +9,6 @@ from .base import (
     AnnotationStoreReader,
     ArrayView,
     FsspecJsonWSIReader,
-    NGFFWSIReader,
     OpenSlideWSIReader,
     TIFFWSIReader,
     TIFFWSIReaderDelegate,
@@ -24,6 +23,7 @@ from .detection import is_dicom, is_ngff, is_tiled_tiff, is_url, is_zarr
 # Import reader classes
 from .dicom import DICOMWSIReader
 from .jp2 import JP2WSIReader
+from .ngff import NGFFWSIReader
 
 __all__ = [
     "AnnotationStoreReader",
